Add packaging support and build scripts
