+++ pyproject.toml
@@ -1,0 +1,15 @@
+[build-system]
+requires = ["setuptools>=68", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "smtp-burst"
+version = "0.1.0"
+description = "Send bursts of SMTP emails for testing throughput"
+authors = [{name = "Example", email = "example@example.com"}]
+readme = "readme.md"
+requires-python = ">=3.11"
+license = {file = "license.md"}
+
+[project.scripts]
+smtp-burst = "smtpburst.__main__:main"
