+++ scripts/build_ubuntu.sh
@@ -1,0 +1,3 @@
+#!/usr/bin/env bash
+set -e
+pyinstaller --onefile -n smtp-burst smtpburst/__main__.py --distpath dist/ubuntu
