+++ scripts/build_windows.bat
@@ -1,0 +1,4 @@
+@echo off
+REM Build Windows executable using PyInstaller
+pyinstaller --onefile -n smtp-burst smtpburst\__main__.py --distpath dist\windows
+if %errorlevel% neq 0 exit /b %errorlevel%
